+++ blueprints/patient_routes.py
@@ -1,0 +1,7 @@
+from flask import Blueprint, render_template
+
+patient_bp = Blueprint('patient', __name__, template_folder='../templates')
+
+@patient_bp.route("/patient", methods=["GET"])
+def patient():
+    return render_template("patient.html")
